# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TERRIA API",
    "summary": "Backend for the shareable history of a field.",
    "description": "Initial CRUD API for fields, public field passports, and timelapse engine.",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/fields/{field_id}/timelapses": {
      "post": {
        "tags": [
          "Timelapse"
        ],
        "summary": "Create or queue a timelapse dataset generation job",
        "operationId": "create_or_get_timelapse_v1_fields__field_id__timelapses_post",
        "parameters": [
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Field Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimelapseJobCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/TimelapseJobResponse"
                    },
                    {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  ],
                  "title": "Response Create Or Get Timelapse V1 Fields  Field Id  Timelapses Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Timelapse"
        ],
        "summary": "List generated timelapse datasets for a field",
        "operationId": "list_field_timelapses_v1_fields__field_id__timelapses_get",
        "parameters": [
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Field Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TimelapseDatasetSummary"
                  },
                  "title": "Response List Field Timelapses V1 Fields  Field Id  Timelapses Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/timelapse-jobs/{job_id}": {
      "get": {
        "tags": [
          "Timelapse"
        ],
        "summary": "Get status of a timelapse generation job",
        "operationId": "get_timelapse_job_v1_timelapse_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelapseJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/fields/{field_id}/timelapses/{dataset_id}": {
      "get": {
        "tags": [
          "Timelapse"
        ],
        "summary": "Get private timelapse manifest",
        "operationId": "get_timelapse_manifest_v1_fields__field_id__timelapses__dataset_id__get",
        "parameters": [
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Field Id"
            }
          },
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelapseManifest"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/fields/{field_id}/timelapses/{dataset_id}/frames/{frame_id}": {
      "get": {
        "tags": [
          "Timelapse"
        ],
        "summary": "Get details and asset URLs for a specific frame",
        "operationId": "get_timelapse_frame_v1_fields__field_id__timelapses__dataset_id__frames__frame_id__get",
        "parameters": [
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Field Id"
            }
          },
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dataset Id"
            }
          },
          {
            "name": "frame_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Frame Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelapseFrame"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/fields/{field_id}/timelapses/{dataset_id}/frames/{frame_id}/assets/{layer}": {
      "get": {
        "tags": [
          "Timelapse"
        ],
        "summary": "Download or view RGB / NDVI asset image for a frame",
        "operationId": "get_timelapse_frame_asset_v1_fields__field_id__timelapses__dataset_id__frames__frame_id__assets__layer__get",
        "parameters": [
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Field Id"
            }
          },
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dataset Id"
            }
          },
          {
            "name": "frame_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Frame Id"
            }
          },
          {
            "name": "layer",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "rgb",
                "ndvi"
              ],
              "type": "string",
              "title": "Layer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/fields/{field_id}/timelapses/{dataset_id}/timeline-state": {
      "get": {
        "tags": [
          "Timelapse"
        ],
        "summary": "Get resolved timeline state for a given date",
        "operationId": "get_timeline_state_v1_fields__field_id__timelapses__dataset_id__timeline_state_get",
        "parameters": [
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Field Id"
            }
          },
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dataset Id"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/fields/{field_id}/timelapses/{dataset_id}/publish": {
      "post": {
        "tags": [
          "Timelapse"
        ],
        "summary": "Publish timelapse dataset for the public field passport",
        "operationId": "publish_timelapse_v1_fields__field_id__timelapses__dataset_id__publish_post",
        "parameters": [
          {
            "name": "field_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Field Id"
            }
          },
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dataset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Publish Timelapse V1 Fields  Field Id  Timelapses  Dataset Id  Publish Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/public/fields/{public_slug}/timelapse": {
      "get": {
        "tags": [
          "Public fields"
        ],
        "summary": "Get public timelapse manifest for a published field",
        "operationId": "get_public_timelapse_v1_public_fields__public_slug__timelapse_get",
        "parameters": [
          {
            "name": "public_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Public Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicTimelapseManifest"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/public/fields/{public_slug}/timelapse/frames/{frame_id}/assets/{layer}": {
      "get": {
        "tags": [
          "Public fields"
        ],
        "summary": "Get asset image for a published field frame",
        "operationId": "get_public_timelapse_asset_v1_public_fields__public_slug__timelapse_frames__frame_id__assets__layer__get",
        "parameters": [
          {
            "name": "public_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Public Slug"
            }
          },
          {
            "name": "frame_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Frame Id"
            }
          },
          {
            "name": "layer",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "rgb",
                "ndvi"
              ],
              "type": "string",
              "title": "Layer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TimelapseDatasetSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "field_id": {
            "type": "string",
            "format": "uuid",
            "title": "Field Id"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "is_demo": {
            "type": "boolean",
            "title": "Is Demo"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "frames_count": {
            "type": "integer",
            "title": "Frames Count"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "field_id",
          "start_date",
          "end_date",
          "status",
          "is_demo",
          "generated_at",
          "frames_count"
        ],
        "title": "TimelapseDatasetSummary"
      },
...
      "TimelapseJobCreate": {
        "properties": {
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "layers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Layers",
            "default": [
              "rgb",
              "ndvi"
            ]
          },
          "is_demo": {
            "type": "boolean",
            "title": "Is Demo",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "start_date",
          "end_date"
        ],
        "title": "TimelapseJobCreate"
      },
      "TimelapseJobResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "field_id": {
            "type": "string",
            "format": "uuid",
            "title": "Field Id"
          },
          "geometry_version_id": {
            "type": "string",
            "format": "uuid",
            "title": "Geometry Version Id"
          },
          "request_hash": {
            "type": "string",
            "title": "Request Hash"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "processing",
              "ready",
              "partial",
              "failed"
            ],
            "title": "Status"
          },
          "progress": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Progress"
          },
          "dataset_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Id"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "field_id",
          "geometry_version_id",
          "request_hash",
          "status",
          "progress",
          "created_at",
          "updated_at"
        ],
        "title": "TimelapseJobResponse"
      },
      "TimelapseManifest": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dataset Id"
          },
          "schema_version": {
            "type": "string",
            "const": "1",
            "title": "Schema Version",
            "default": "1"
          },
          "processing_version": {
            "type": "string",
            "title": "Processing Version"
          },
          "field_id": {
            "type": "string",
            "format": "uuid",
            "title": "Field Id"
          },
          "geometry_version_id": {
            "type": "string",
            "format": "uuid",
            "title": "Geometry Version Id"
          },
          "boundary": {
            "$ref": "#/components/schemas/PolygonGeometry"
          },
          "area_hectares": {
            "type": "number",
            "title": "Area Hectares"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "default": "America/Argentina/Cordoba"
          },
          "status": {
            "type": "string",
            "enum": [
              "ready",
              "partial",
              "failed"
            ],
            "title": "Status"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "is_demo": {
            "type": "boolean",
            "title": "Is Demo",
            "default": false
          },
          "playback": {
            "$ref": "#/components/schemas/PlaybackConfig"
          },
          "frames": {
            "items": {
              "$ref": "#/components/schemas/TimelapseFrame"
            },
            "type": "array",
            "title": "Frames",
            "default": []
          },
          "weather_daily": {
            "items": {
              "$ref": "#/components/schemas/WeatherDaily"
            },
            "type": "array",
            "title": "Weather Daily",
            "default": []
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/TimelapseSource"
            },
            "type": "array",
            "title": "Sources",
            "default": []
          },
          "missing_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Reasons",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "dataset_id",
          "processing_version",
          "field_id",
          "geometry_version_id",
          "boundary",
          "area_hectares",
          "start_date",
          "end_date",
          "status",
          "generated_at"
        ],
        "title": "TimelapseManifest"
      },
...
    }
  },
...
}